chore: move non-deployment operations to a new script
