+++ deploy/007_deploy_sbrick.js
@@ -1,26 +1,14 @@
-const { ethers } = require('hardhat');
-
 module.exports = async ({ getNamedAccounts, deployments }) => {
-  const { deploy, get } = deployments;
+  const { deploy } = deployments;
   const { deployer } = await getNamedAccounts();
 
   const chainId = await getChainId();
   const { name, symbol } = config.protocolParameters[chainId].sbrick;
 
-  const deployment = await deploy('sOlympus', {
+  await deploy('sOlympus', {
     from: deployer,
     args: [name, symbol],
     log: true,
   });
-
-  const treasuryAddress = (await get('OlympusTreasury')).address;
-  const treasury = (
-    await ethers.getContractFactory('OlympusTreasury')
-  ).attach(treasuryAddress);
-  // NOTE: set sBRICK
-  const sBrickQueue = await treasury.sOHMQueue();
-  if (sBrickQueue.eq(0)) {
-    await treasury.queue('9', deployment.address);
-  }
 };
 module.exports.tags = ['sBRICK', 'AllEnvironments'];
